+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -230,7 +230,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5QVJ9JZ6ZN3537D17CV5MZV"
+        "01K5R6RX9X7EPNP3XM3XG59V07"
     )
     r"""The id of the resource"""