+++ packages/orq-rc/src/orq_ai_sdk/models/listbudgetsop.py
@@ -242,7 +242,7 @@ class ListBudgetsData(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-22T04:40:19.773Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-22T07:56:11.042Z")
     r"""The date and time the resource was last updated"""