+++ packages/orq-rc/src/orq_ai_sdk/models/listcontactsop.py
@@ -184,7 +184,7 @@ class Data(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-22T04:40:19.773Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-22T07:56:11.042Z")
     r"""The date and time the resource was last updated"""
 
     @model_serializer(mode="wrap")