+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -132,7 +132,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01K5QVJ9JX77R0PA4VY087RSR8"
+        "01K5R6RX9WYRS0Q60CFN2MY0XS"
     )
     r"""The id of the resource"""