+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasetop.py
@@ -119,5 +119,5 @@ class UpdateDatasetResponseBody(BaseModel):
     created: Optional[datetime] = None
     r"""The date and time the resource was created"""
 
-    updated: Optional[datetime] = parse_datetime("2025-09-22T04:40:19.773Z")
+    updated: Optional[datetime] = parse_datetime("2025-09-22T07:56:11.042Z")
     r"""The date and time the resource was last updated"""