+++ packages/orq-rc/RELEASES.md
@@ -2819,3 +2819,13 @@ Based on:
 - [python v3.13.0-rc.22] packages/orq-rc
 ### Releases
 - [PyPI v3.13.0-rc.22] https://pypi.org/project/orq-ai-sdk/3.13.0-rc.22 - packages/orq-rc
+
+## 2025-09-22 07:57:47
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.622.1 (2.709.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v3.13.0-rc.23] packages/orq-rc
+### Releases
+- [PyPI v3.13.0-rc.23] https://pypi.org/project/orq-ai-sdk/3.13.0-rc.23 - packages/orq-rc